refactor: Move CoverageSummary modules into root of components directory
